Added death animation (FOR NOW: PRESS 'T' TO DAMAGE PLAYER)
PRESS 'H' TO HEAL PLAYER.
These controls are debug code that we will remove later, to be used for testing the health system.
The player also dies when the time runs out.
The game currently softlocks on player death. That is, the timer stops animating and the player can no longer move.